Move interactive graph editor components
