+++ packages/perseus-editor/src/components/__tests__/interactive-graph-settings.test.tsx
@@ -4,7 +4,7 @@ import {userEvent as userEventLib} from "@testing-library/user-event";
 import * as React from "react";
 
 import {testDependencies} from "../../../../../testing/test-dependencies";
-import InteractiveGraphSettings from "../interactive-graph-settings";
+import InteractiveGraphSettings from "../../widgets/interactive-graph-editor/components/interactive-graph-settings";
 
 import type {UserEvent} from "@testing-library/user-event";
