+++ packages/perseus-editor/src/widgets/interactive-graph-editor/components/graph-points-count-selector.tsx
@@ -2,9 +2,9 @@ import {OptionItem, SingleSelect} from "@khanacademy/wonder-blocks-dropdown";
 import {StyleSheet} from "aphrodite";
 import * as React from "react";
 
-import {UNLIMITED, parsePointCount} from "../util/points";
+import {UNLIMITED, parsePointCount} from "../../../util/points";
 
-import type {PointValue} from "../util/points";
+import type {PointValue} from "../../../util/points";
 
 const GraphPointsCountSelector = ({
     numPoints = 1,
