+++ packages/perseus-editor/src/widgets/interactive-graph-editor/interactive-graph-editor.tsx
@@ -19,12 +19,12 @@ import _ from "underscore";
 
 import LabeledRow from "../../components/graph-locked-figures/labeled-row";
 import LockedFiguresSection from "../../components/graph-locked-figures/locked-figures-section";
-import GraphPointsCountSelector from "../../components/graph-points-count-selector";
-import GraphTypeSelector from "../../components/graph-type-selector";
-import {InteractiveGraphCorrectAnswer} from "../../components/interactive-graph-correct-answer";
+import GraphPointsCountSelector from "./components/graph-points-count-selector";
+import GraphTypeSelector from "./components/graph-type-selector";
+import {InteractiveGraphCorrectAnswer} from "./components/interactive-graph-correct-answer";
 import InteractiveGraphDescription from "../../components/interactive-graph-description";
-import InteractiveGraphSettings from "../../components/interactive-graph-settings";
-import SegmentCountSelector from "../../components/segment-count-selector";
+import InteractiveGraphSettings from "./components/interactive-graph-settings";
+import SegmentCountSelector from "./components/segment-count-selector";
 import StartCoordsSettings from "./start-coords/start-coords-settings";
 import {shouldShowStartCoordsUI} from "../../components/util";
 import {parsePointCount} from "../../util/points";
